+++ drizzle/0015_naive_tinkerer.sql
@@ -1,0 +1,3 @@
+ALTER TABLE `sources` ADD `expires_at` integer;--> statement-breakpoint
+ALTER TABLE `sources` ADD `max_connections` integer;--> statement-breakpoint
+ALTER TABLE `sources` ADD `account_status` text;